+++ src/components/pages/campaigns-create-approve/components/erc-1155/index.tsx
@@ -19,7 +19,7 @@ import LinksContents from '../links-contents'
 import { RootState, IAppDispatch } from 'data/store'
 import { connect } from 'react-redux'
 import { useParams } from 'react-router-dom'
-import { TTokenType, TLinkContent, TAlchemyNFTToken } from 'types'
+import { TTokenType, TLinkContent, TAlchemyNFTToken, TClaimPattern } from 'types'
 import {
   WidgetComponent
 } from 'components/pages/common'
@@ -138,7 +138,8 @@ const createSelectContainer = (
   tokenAddress: string | null,
   userAddress: string,
   signer: any,
-  checkIfDisabled: () => boolean
+  checkIfDisabled: () => boolean,
+  claimPattern: TClaimPattern
 ) => {
   return <InputsContainer>
     <SelectStyled
@@ -150,8 +151,20 @@ const createSelectContainer = (
           return alertError(`Token #${tokenId} was already added`)
         }
 
+        // if "not found" was clicked
         if (typeof value === 'string') {
-          // if "not found" was clicked
+          if (claimPattern === 'mint') {
+            return setAssetsData([
+              ...assetsData, {
+                tokenId: value,
+                tokenAmount: "1",
+                linksAmount: "1",
+                type: 'ERC1155',
+                id: assetsData.length,
+                tokenName: 'Token ERC1155'
+              }
+            ])
+          }
           const userOwnership = await defineIfUserOwnsToken(
             userAddress,
             'ERC1155',
@@ -210,7 +223,8 @@ const createTextInputOrSelect = (
   nfts: TAlchemyNFTToken[],
   tokenAddress: string | null,
   userAddress: string,
-  signer: any
+  signer: any,
+  claimPattern: TClaimPattern
 ) => {
 
   if (enabledInput) {
@@ -233,7 +247,8 @@ const createTextInputOrSelect = (
     tokenAddress,
     userAddress,
     signer,
-    checkIfDisabled
+    checkIfDisabled,
+    claimPattern
   )
 
 }
@@ -378,7 +393,8 @@ const Erc1155: FC<ReduxType > = ({
         nfts,
         tokenAddress,
         address,
-        signer
+        signer,
+        claimPattern
       )}
       <LinksContents
         type={type}