+++ src/components/pages/collection/styled-components.tsx
@@ -2,7 +2,9 @@ import styled from 'styled-components'
 import {
   Aside,
   TableValue,
-  WidgetComponent
+  WidgetComponent,
+  WidgetTitle,
+  AttentionContainer
 } from 'components/pages/common'
 import { Button } from 'components/common'
 
@@ -31,7 +33,7 @@ export const AsideContent = styled.div`
 
 export const MainContent = styled.div`
   margin-right: 24px;
-  flex: 1 0 50%; 
+  flex: 1 0; 
 `
 
 export const TableValueFlex = styled(TableValue)`
@@ -53,17 +55,42 @@ export const ButtonFullWidth = styled(Button)`
 `
 
 export const TokensList = styled.ul`
-  max-width: 652px;
   padding: 0;
-  margin: 0 0 32px;
+  margin: 0;
   list-style: none;
-  border-bottom: 1px solid ${props => props.theme.additionalBorderColor};
 `
 
 export const TokenItem = styled.li`
   margin-bottom: 32px;
+  
+  &:last-child {
+    margin-bottom: 0;
+  }
 `
 
 export const CopyIcon = styled.span`
   cursor: pointer;
+`
+
+export const Header = styled.div`
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 32px;
+`
+
+export const WidgetTitleStyled = styled(WidgetTitle)`
+  max-width: 400px;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  margin-bottom: 0;
+`
+
+export const ContainerButton = styled(Button)`
+  align-self: end;
+  margin-bottom: 0px;
+`
+
+export const AttentionContainerStyled = styled(AttentionContainer)`
+  margin-bottom: 24px;
 `