+++ src/data/store/reducers/campaign/async-actions/index.tsx
@@ -1,4 +1,3 @@
-import createProxyContract from './create-proxy'
 import setAssetsData from './set-assets-data'
 import setTokenContractData from './set-token-contract-data'
 import generateLinks from './generate-links'
@@ -7,9 +6,10 @@ import applyClaimPattern from './apply-claim-pattern'
 import resetCampaign from './reset-campaign'
 import openDraft from './open-draft'
 import getUserNFTs from './get-user-nfts'
+import createProxyAddress from './create-proxy-address'
 
 export {
-  createProxyContract,
+  createProxyAddress,
   setAssetsData,
   setTokenContractData,
   applyClaimPattern,