+++ src/data/store/reducers/collections/async-actions/index.tsx
@@ -1,9 +1,11 @@
 import createCollectionERC1155 from './create-collection-erc1155'
 import createTokenERC1155 from './create-token-erc1155'
 import getCollectionData from './get-collection-data'
+import createClaimLinks from './create-claim-links'
 
 export {
   createCollectionERC1155,
   createTokenERC1155,
-  getCollectionData
+  getCollectionData,
+  createClaimLinks
 }